+++ app/presenters/facet/collection_presenter.rb
@@ -3,9 +3,27 @@
 module Facet
   class CollectionPresenter < SimplePresenter
     def add_facet_url(item)
-      value = facet_value_for_facet_item(item)
-      base_url = default_facet_value?(value) ? search_url : collection_url(value)
-      [base_url, facet_item_url_base_query].join('?')
+      fail NotImplementedError
+    end
+
+    def remove_facet_path(_item)
+      search_path
+    end
+
+    def replace_facet_path(item)
+      default_facet_value?(item.value) ? search_path : collection_path(item.value)
+    end
+
+    def add_facet_query(item, base: facet_item_url_base_query)
+      fail NotImplementedError
+    end
+
+    def remove_facet_query(_item)
+      facet_item_url_base_query
+    end
+
+    def replace_facet_query(_item)
+      facet_item_url_base_query
     end
 
     def apply_order_to_items?
@@ -16,18 +34,12 @@ module Facet
       items.unshift(items.delete(items.detect { |item| facet_in_params?(facet_name, item) }))
     end
 
-    def remove_facet_url(_item)
-      [search_url, facet_item_url_base_query].reject(&:blank?).join('?')
+    def remove_facet_query(_item)
+      facet_item_url_base_query
     end
 
     def facet_in_params?(field, item)
-      value = facet_value_for_facet_item(item)
-
-      facet_params(field) == value
-    end
-
-    def default_facet_value
-      'all'
+      facet_params(field) == item.value
     end
 
     def facet_params(_field)
